Refactor to feature-based architecture with code quality cleanup
Restructure codebase from flat layers (pages/, components/) to feature-based
(features/lesson/, features/review/, features/path/, features/exercises/) with
shared utilities and components.

- Break up Lesson.tsx (380 lines) into 5 files: LessonPage, useLessonState,
  CompletionScreen, ExitButton, LessonHeader
- Break up Review.tsx (221 lines) into 4 files: ReviewPage, useReviewSession,
  ReviewIntro, ReviewSummary
- Deduplicate shuffle() (3 copies → 1), correctAnswer normalization (4 → 1),
  renderExercise dispatcher (2 → 1)
- Extract shared components: ProgressBar, LoadingScreen, EmptyState, CloseIcon
- Centralize hydration in HydrationGuard (fixes deep-link bug)
- Fix layer violation: move SRS card dedup from Lesson.tsx into srsStore
- Move stray types (UnitStatus, LessonResult, ReviewSession, ReviewResult)
  into types/
- Remove dead code: unused checkAnswer(), lessonsCompleted Dexie table, empty
  hooks/ directory

Closes #26
