+++ frontend/src/engine/reviewRunner.ts
@@ -1,14 +1,6 @@
 import type { Exercise, SRSCard, VocabEntry } from '../types';
 import { db } from '../stores/db';
-
-function shuffle<T>(arr: T[]): T[] {
-  const a = [...arr];
-  for (let i = a.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [a[i], a[j]] = [a[j], a[i]];
-  }
-  return a;
-}
+import { shuffle } from '../shared/utils/shuffle';
 
 /**
  * Pick up to `count` random distractors from the vocabulary table,
