+++ frontend/src/features/path/UnitCard.tsx
@@ -1,7 +1,6 @@
-import type { Unit, LessonScore } from '@/types';
+import type { Unit, LessonScore, UnitStatus } from '@/types';
 import LessonList from './LessonList';
-
-export type UnitStatus = 'locked' | 'available' | 'in_progress' | 'completed';
+import StatusIcon from './StatusIcon';
 
 interface UnitCardProps {
   unit: Unit;
@@ -89,43 +88,4 @@ export default function UnitCard({
       )}
     </div>
   );
-}
-
-function StatusIcon({ status }: { status: UnitStatus }) {
-  const base = 'w-8 h-8 rounded-full flex items-center justify-center flex-shrink-0';
-
-  switch (status) {
-    case 'locked':
-      return (
-        <div className={`${base} bg-gray-200`}>
-          <svg className="w-4 h-4 text-gray-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-            <path strokeLinecap="round" strokeLinejoin="round" d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z" />
-          </svg>
-        </div>
-      );
-    case 'available':
-      return (
-        <div className={`${base} bg-blue-100`}>
-          <svg className="w-4 h-4 text-blue-600" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-            <path strokeLinecap="round" strokeLinejoin="round" d="M14.752 11.168l-3.197-2.132A1 1 0 0010 9.87v4.263a1 1 0 001.555.832l3.197-2.132a1 1 0 000-1.664z" />
-          </svg>
-        </div>
-      );
-    case 'in_progress':
-      return (
-        <div className={`${base} bg-blue-100`}>
-          <svg className="w-4 h-4 text-blue-600" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-            <path strokeLinecap="round" strokeLinejoin="round" d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
-          </svg>
-        </div>
-      );
-    case 'completed':
-      return (
-        <div className={`${base} bg-green-100`}>
-          <svg className="w-4 h-4 text-green-600" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-            <path strokeLinecap="round" strokeLinejoin="round" d="M5 13l4 4L19 7" />
-          </svg>
-        </div>
-      );
-  }
 }
