+++ frontend/src/shared/components/LoadingScreen.tsx
@@ -1,0 +1,7 @@
+export default function LoadingScreen() {
+  return (
+    <div className="min-h-screen bg-gray-50 flex items-center justify-center">
+      <p className="text-gray-400">Loading...</p>
+    </div>
+  );
+}
