+++ frontend/src/shared/components/ProgressBar.tsx
@@ -1,0 +1,14 @@
+interface ProgressBarProps {
+  progress: number;
+}
+
+export default function ProgressBar({ progress }: ProgressBarProps) {
+  return (
+    <div className="h-2 bg-gray-200 rounded-full overflow-hidden">
+      <div
+        className="h-full bg-blue-500 rounded-full transition-all duration-300"
+        style={{ width: `${progress}%` }}
+      />
+    </div>
+  );
+}
