+++ frontend/src/shared/utils/shuffle.ts
@@ -1,0 +1,9 @@
+/** Fisher-Yates shuffle. Returns a new array. */
+export function shuffle<T>(arr: T[]): T[] {
+  const a = [...arr];
+  for (let i = a.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [a[i], a[j]] = [a[j], a[i]];
+  }
+  return a;
+}
